src/main/java/com/caglar/todoapp/impl/UserService.java
package com.caglar.todoapp.impl;

import com.caglar.todoapp.entity.User;
import com.caglar.todoapp.repository.UserRepository;
import com.caglar.todoapp.service.IUserService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Service;

@Service
@Slf4j
public class UserService implements IUserService {

    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public User saveUser(User user) {
        User userSaved = userRepository.save(user);
        log.debug("User is saved with userId={}", userSaved.getId());
        return user;
    }

    @Override
    public User findUserByEmail(String email) {
        return userRepository.findByEmail(email).orElseThrow();
    }

    @Override
    public User getCurrentUser() {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        return (User) authentication.getPrincipal();
    }
}
